refactor(lint): rename to author_lifecycle, harden chain validation (#256)
Round 3 review:

[critical] Active authors were treated as clean without any signature
or body-integrity verification. The module name 'signature' implied
crypto verification we cannot do at P0 (Ed25519 deferred per
verifier.rs:8-22), creating false operator confidence that 'no
finding' meant 'signature verified'.

  Fix: rename module signature.rs -> author_lifecycle.rs and fn
  check_signature -> check_author_lifecycle. Module doc has an
  explicit 'What this is *not*' section: returning None does NOT
  imply record.signature is valid, body untampered, or key version
  current — only that the author identity is Active and the chain
  shape passes domain validation. Real Ed25519 + body-integrity
  remain P1 follow-ups.

[high] Malformed-chain detection only caught missing-Author. A
persisted record with two Authors or out-of-order roles (Sensor
before Author, etc.) could pass on .find() returning the first
Author. Relying on ingest-time validation is unsafe under direct
DB tamper or partial migration.

  Fix: re-run validate_chain on every call. Any §4.2 ordering /
  role-count violation now emits ChainStatus::Malformed with author=None.
  New tests cover duplicate-Author and Sensor-before-Author cases.
